Update ex00 ex01 and added ex02
update ex00 and ex01. Added ex02 of this assignment which is we can't need to do that to pass this module. But it has some nice information about vectors iterations etc. I believe it has to be at least take a look before move on to learn some good stuff in c++.